lab567/requirements.txt
aiohttp==3.6.2
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.1
azure-core==1.5.0
azure-eventhub==5.1.0
azure-eventhub-checkpointstoreblob-aio==1.1.0
bcrypt==3.1.6
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu4
docopt==0.6.2
duplicity==0.8.4
entrypoints==0.3
environs==7.4.0
fasteners==0.12.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
httplib2==0.11.3
idna==2.6
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.6.0
monotonic==1.5
msrest==0.6.13
multidict==4.7.5
munch==2.5.0
netifaces==0.10.4
num2words==0.5.10
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.1.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.7.0
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.3
python-dateutil==2.7.3
python-debian==0.1.36
python-dotenv==0.13.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
redis==3.5.1
reportlab==3.5.23
requests==2.21.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
system-service==0.3
systemd-python==234
uamqp==1.2.7
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.7
vboxapi==1.0
virtualenv==15.1.0
wadllib==1.3.3
waitress==1.4.3
Werkzeug==1.0.1
WTForms==2.2.1
xkit==0.0.0
yarl==1.4.2
